requirements.txt
telegram
python-telegram-bot
scipy
geopy
folium
pillow